Moved post_processing and error_analysis into a post processing module
